+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - fix D10 deprecations: Creation of dynamic property is deprecated
 
+### Added
+- add event PRE_PROCESS_FILE to allow client or file alteration before Tika OCR
+
 ## [1.0.0] - 2023-01-27
 ### Added
 - init module